Refactor to fall in line with PR feedback
Generally make everything going on more explicit and improve
naming
