+++ app/controllers/candidate_interface/apply_from_find_controller.rb
@@ -7,9 +7,13 @@ module CandidateInterface
     rescue_from ActionController::ParameterMissing, with: :render_not_found
 
     def show
-      service = ValidateCourseQueryStringParams.new(provider_code: params[:providerCode], course_code: params[:courseCode])
-      service.execute
-      @course = service.return_course
+      service = ApplyFromFindPage.new(provider_code: params[:providerCode],
+                                      course_code: params[:courseCode],
+                                      can_apply_on_apply: false,
+                                      course_on_find: false,
+                                      course: nil)
+      service.determine_whether_course_is_on_find_or_apply
+      @course = service.course
 
       if service.can_apply_on_apply?
         render :apply_on_ucas_or_apply
